+++ src/Writer/JobWriter.php
@@ -25,8 +25,6 @@ declare(strict_types=1);
 
 namespace Teknoo\East\Paas\Writer;
 
-use Teknoo\East\Common\Contracts\DBSource\ManagerInterface;
-use Teknoo\East\Common\Service\DatesService;
 use Teknoo\Recipe\Promise\PromiseInterface;
 use Teknoo\East\Paas\Object\Job;
 use Teknoo\East\Common\Contracts\Object\ObjectInterface;
@@ -52,13 +50,6 @@ class JobWriter implements WriterInterface
      * @use PersistTrait<Job>
      */
     use PersistTrait;
-
-    public function __construct(
-        private ManagerInterface $manager,
-        private ?DatesService $datesService = null,
-        protected bool $prefereRealDateOnUpdate = true,
-    ) {
-    }
 
     public function save(
         ObjectInterface $object,